setplanet.c compile fix
      * setplanet.c: isalpha requires int not char.  Similar problem seen in
        calls to toupper and isspace.
M ./Vanilla/tools/setplanet.c -1 +1

+++ Vanilla/tools/setplanet.c
@@ -194,7 +194,7 @@ int main(int argc, char **argv)
     }
 
     /* check for planet identifier */
-    if (isalpha(argv[i][0])) {
+    if (isalpha((int)argv[i][0])) {
       pl = planet_find(argv[i]);
     } else {
       pl = planet_by_number(argv[i]);
